pythonlearning-ashokpractise/SR-HACKATHON3/requirements.txt
django
djangorestframework
pymysql
django-cors-headers
